Move goto_symext::symex_assign to goto_symex.cpp
symex_assign.cpp now only contains functions that are not members of
goto_symext.
This will make the separation between members and non-members clearer.
