+++ src/goto-symex/symex_assign.cpp
@@ -10,17 +10,10 @@ Author: Daniel Kroening, kroening@kroening.com
 /// Symbolic Execution
 
 #include "goto_symex.h"
-
+#include "goto_symex_state.h"
 #include <util/byte_operators.h>
-#include <util/c_types.h>
-#include <util/cprover_prefix.h>
-#include <util/exception_utils.h>
 #include <util/expr_util.h>
 #include <util/format_expr.h>
-#include <util/pointer_offset_size.h>
-#include <util/simplify_expr.h>
-
-#include "goto_symex_state.h"
 
 // We can either use with_exprt or update_exprt when building expressions that
 // modify components of an array or a struct. Set USE_UPDATE to use
@@ -92,69 +85,6 @@ static void symex_assign_byte_extract(
   const namespacet &ns,
   const symex_configt &symex_config,
   symex_targett &target);
-
-void goto_symext::symex_assign(statet &state, const code_assignt &code)
-{
-  exprt lhs = clean_expr(code.lhs(), state, true);
-  exprt rhs = clean_expr(code.rhs(), state, false);
-
-  DATA_INVARIANT(
-    lhs.type() == rhs.type(), "assignments must be type consistent");
-
-  log.conditional_output(
-    log.debug(), [this, &lhs](messaget::mstreamt &mstream) {
-      mstream << "Assignment to " << format(lhs) << " ["
-              << pointer_offset_bits(lhs.type(), ns).value_or(0) << " bits]"
-              << messaget::eom;
-    });
-
-  if(rhs.id()==ID_side_effect)
-  {
-    const side_effect_exprt &side_effect_expr=to_side_effect_expr(rhs);
-    const irep_idt &statement=side_effect_expr.get_statement();
-
-    if(
-      statement == ID_cpp_new || statement == ID_cpp_new_array ||
-      statement == ID_java_new_array_data)
-      symex_cpp_new(state, lhs, side_effect_expr);
-    else if(statement==ID_allocate)
-      symex_allocate(state, lhs, side_effect_expr);
-    else if(statement == ID_va_start)
-      symex_va_start(state, lhs, side_effect_expr);
-    else
-      UNREACHABLE;
-  }
-  else
-  {
-    assignment_typet assignment_type=symex_targett::assignment_typet::STATE;
-
-    // Let's hide return value assignments.
-    if(lhs.id()==ID_symbol &&
-       id2string(to_symbol_expr(lhs).get_identifier()).find(
-                  "#return_value!")!=std::string::npos)
-      assignment_type=symex_targett::assignment_typet::HIDDEN;
-
-    // We hide if we are in a hidden function.
-    if(state.call_stack().top().hidden_function)
-      assignment_type=symex_targett::assignment_typet::HIDDEN;
-
-    // We hide if we are executing a hidden instruction.
-    if(state.source.pc->source_location.get_hide())
-      assignment_type=symex_targett::assignment_typet::HIDDEN;
-
-    exprt::operandst lhs_if_then_else_conditions;
-    symex_assign_rec(
-      state,
-      lhs,
-      nil_exprt(),
-      rhs,
-      lhs_if_then_else_conditions,
-      assignment_type,
-      ns,
-      symex_config,
-      target);
-  }
-}
 
 /// Store the \p what expression by recursively descending into the operands
 /// of \p lhs until the first operand \c op0 is _nil_: this _nil_ operand
